Implement sendToContractTransparentCircuit
