+++ contracts/transfer/circuits/Cargo.toml
@@ -5,17 +5,16 @@ authors = ["Bounce23 <luke@dusk.network>", "Victor Lopez <victor@dusk.network>"]
 edition = "2018"
 
 [dependencies]
-rand = "0.7"
 rand_core = "0.5"
 anyhow = "1.0"
-dusk-plonk = "0.3.6"
+dusk-plonk = "0.3"
 canonical = "0.4"
-canonical_derive = "0.4"
 phoenix-core = {git="https://github.com/dusk-network/phoenix-core", branch="vlopes11/note-refactor-blinding-factor", features = ["canon"]}
 poseidon252 = {git="https://github.com/dusk-network/Poseidon252", tag="v0.15.0", features=["canon"]}
-plonk_gadgets = {git="https://github.com/dusk-network/plonk_gadgets", tag="v0.4.3"}
 dusk-pki = {git="https://github.com/dusk-network/dusk-pki", tag="v0.4.1"}
 schnorr = {git="https://github.com/dusk-network/schnorr", tag="v0.2.0"}
 
 [dev-dependencies]
 canonical_host = "0.4"
+canonical_derive = "0.4"
+rand = "0.7"
