+++ contracts/transfer/circuits/src/tests.rs
@@ -1,0 +1,1 @@
+pub mod leaf;
